Session Expiration managed

+++ src/Component/Client/Createtask.js
@@ -1,13 +1,42 @@
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import axios from 'axios';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { ToastContainer, toast } from 'react-toastify';
+import { logout } from '../../Services/Actions/Authenticationaction';
+import { useNavigate } from 'react-router';
 const Createtask = () => {
     const [title, setTitle] = useState("");
     const [description, setDescription] = useState("");
     const [file, setFile] = useState(null);
+    var authuser = useSelector((state) => state.auth.user);
+    const dispatch = useDispatch();
+    const navigate = useNavigate();
+    var temp = 0;
 
-    const authuser = useSelector((state) => state.auth.user);
+    useEffect(()=>{
+        if(temp<1){
+            checkToken();
+            temp++;
+        }
+    },[]);
+
+    const checkToken = async () => {
+        await axios.get("http://localhost:8080/login/checkToken",{
+            headers: {
+                'Authorization': `Bearer ${authuser?.token}`
+            }
+        }).then((response) => {
+            console.log(response)
+            // navigate(-1);
+            return;
+        } ).catch((error) => {
+            console.log("in catch "+error.response?.data);
+            const err = error.response?.data;
+            dispatch(logout());
+            navigate("/",{state:{err}});
+            
+        });
+    }
 
     const handleFileChange = (event) => {
         setFile(event.target.files[0]);
@@ -23,47 +52,63 @@ const Createtask = () => {
             status: "NEW",
             clientId: authuser.id
         }
+        if (title !== "" && description !== "") {
+            createTask(task);
+        }
+        else {
+            toast.error("Title and Description is required");
+        }
+    }
+
+
+    const createTask = async (task) => {
         await axios.post("http://localhost:8080/task", task, {
             headers: {
                 'Authorization': `Bearer ${authuser.token}`
             }
         }).then(async (response) => {
-
-            const taskId = response.data?.taskId;
-            const formData = new FormData();
-            formData.append('file', file);
-
-            await axios.post(`http://localhost:8080/document/upload/${taskId}`, formData, {
-                headers: {
-                    'Authorization': `Bearer ${authuser.token}`
-                }
-            }).then((response) => {
-                console.log("done");
-                toast.success("Task is Created Successfully ");
-                clearAll();
-            },
-            async (error) => {
-                clearAll();
-                await axios.delete(`http://localhost:8080/task/${taskId}`, {
+            if (file) {
+                const taskId = response.data?.taskId;
+                const formData = new FormData();
+                formData.append('file', file);
+                await axios.post(`http://localhost:8080/document/upload/${taskId}`, formData, {
                     headers: {
                         'Authorization': `Bearer ${authuser.token}`
                     }
-                })
-                console.log("File uploading fail");
-                toast.error("Max file size is 10Mb! Try Again");
-                
-            })    
+                }).then((response) => {
+                    console.log("done");
+                    toast.success("Task is Created Successfully ");
+                    clearAll();
+                },
+                    async (error) => {
+                        clearAll();
+                        await axios.delete(`http://localhost:8080/task/${taskId}`, {
+                            headers: {
+                                'Authorization': `Bearer ${authuser.token}`
+                            }
+                        })
+                        console.log("File uploading fail");
+                        toast.error("Max file size is 10Mb! Try Again");
+                        console.log(error);
+
+                    })
+            }
+            else { 
+                toast.success("Task is Created Successfully "); 
+                clearAll();
+            }
+
         },
-        (error)=>{
-            toast.error("Something Went Wrong! Try Again");
-            clearAll();
-        })
+            (error) => {
+                toast.error("Something Went Wrong! Try Again");
+                clearAll();
+            })
     }
 
-    const clearAll = ()=>{
+    const clearAll = () => {
         setTitle("");
         setDescription("");
-        setFile("");
+        setFile(Date.now());
     }
 
     return (
